invoke HistoryRecordingStateChangeObserver#onShouldStartRecording with current state
