+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/HistoryRecordingStateHandler.kt
@@ -25,6 +25,9 @@ internal class HistoryRecordingStateHandler : TripSessionStateObserver {
 
     fun registerStateChangeObserver(observer: HistoryRecordingStateChangeObserver) {
         historyRecordingStateChangeObservers.add(observer)
+        if (shouldNotifyOnStart(currentState)) {
+            observer.onShouldStartRecording(currentState)
+        }
     }
 
     fun unregisterStateChangeObserver(observer: HistoryRecordingStateChangeObserver) {
@@ -109,10 +112,14 @@ internal class HistoryRecordingStateHandler : TripSessionStateObserver {
                     finishRecordingBlock(it, oldState)
                 }
             }
-            if (newState !is HistoryRecordingSessionState.Idle) {
+            if (shouldNotifyOnStart(newState)) {
                 historyRecordingStateChangeObservers.forEach { it.onShouldStartRecording(newState) }
             }
             copilotSessionObservers.forEach { it.onCopilotSessionChanged(newState) }
         }
+    }
+
+    private fun shouldNotifyOnStart(state: HistoryRecordingSessionState): Boolean {
+        return state !is HistoryRecordingSessionState.Idle
     }
 }
